eval: fix isTrivial for letExpr's
Value::isTrivial is used in a variety of cases, and mostly
descends throughout thunks. I.e., isTrivial returns true if a
value is a partially evaluated thunk that returns a 'trivially
trivial' value. This patch allows isTrivial to descend into let
expressions, so that if let expressions return a trivial value,
they're treated as such.

Fixes #6753.